Move DOM tests to separate file and use testsets
